requirements.txt
chainlit
langchain
langchain-community
langchain-google-genai
python-dotenv
langchain_experimental
langchain_chroma
langchainhub
pypdf
sentence_transformers
openai
tiktoken
tokenizers
faiss-cpu
unstructured
numpy==1.24.4
nltk==3.9.1
tensorboard
